dmsetup: Add dm package for device mapper to wrap Thin volumes
Borrows heavily from the pre-CRD code, as well as the preliminary
deviceswitch CRD, but isolated into a stand-alone utility package.
The lifecycle is simple enough, with idempotent commands, that even
though we need two layers of device-mapper wrappers per Thin Volume,
we can probably manage it without a full-blown k8s CRD.  This patch
just implements the utility functions, the next patch will put them to
work.

There are some quirks based on whether udev is running in the host and
whether the dm device is linear or multipath.  This was tested with
both device-mapper 1.02.183 in minikube, and device-mapper 1.02.199 in
kcli and Fedora 40.  The patch includes tweaks to the commands package
to match actual error messages in the various situations we require to
be idempotent.

Signed-off-by: Eric Blake <eblake@redhat.com>