crd: Drop DeviceSwitch CRD code
This CRD was envisioned when we first started adding CRDs, but the
more we have refined our design, the less we think we need the
overhead of a k8s CR tracking whether a given node is using NBD or the
LV, and can instead track that directly as part of the Volume CRD plus
information gleaned directly from the node filesystem.  The code for
handling device-mapper wrappers is already incorporated in Volume, and
the remaining code for handling NBD clients will be coming soon.

(Should deletion be premature, we can always revert this patch)

Signed-off-by: Eric Blake <eblake@redhat.com>