deploy: Drop nbd-server ServiceAccount
The qemu-storage-daemon container runs as part of the
node-controller-manager DaemonSet's Pod, which uses the
node-controller-manager ServiceAccount. There are no remaining users of
the nbd-server ServiceAccount.

Note that the qemu-storage-daemon container doesn't need k8s API server
access, but there is no way to isolate it given its privilege
requirements and location in the node-controller-manager Pod. Maybe in
the future it can be separated for better security.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>