volume: Wait for volume activation with WatchPending
The Volume Node reconciler is already watching for changes to
ThinPoolLv, so during a node stage operation, a reconcile may happen
more than once (on the first pass, we update ThinPoolLv but don't see
it active, so updateStatusAttachedToNodes changes nothing; on the
second pass, we have nothing to change in ThinPoolLv but can now
update our status to match the changed status that triggered our
second reconcile).  But this is not as clean as having
reconcileThinAttaching do the waiting, and do it by use of the
WatchPending helper error, especially since that will make the next
patch to add in device-mapper wrappers easier.

Signed-off-by: Eric Blake <eblake@redhat.com>