csi: Publish Filesystem volumes using bind mounts
Kubelet recursively changes gids on volumes when the Pod SecurityContext
fsGroup setting is specified. This doesn't work as expected on KubeSAN's
symlinks to the mounted staging directory. Kubelet changes the gid on
the symlink itself and does not recurse into the Filesystem volume's
contents.

Use a bind mount when publishing Filesystem volumes so that Kubelet can
change gids correctly when fsGroup is specified.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>