Makefile, Containerfile: run makefile as part of container build
The CI Pipeline was broken because it didn't have native go installed
when it ran the Makefile, which uses go to build artifacts. This patch
runs the Makefile as part of the container build and uses the Container's
go version for consistency. The containerized-build script runs the
commands to build everything in the Container.

It also fixes a bug in the Makefile where the controller-gen target
wasn't running.

Signed-off-by: Chandan Somani <csomani@redhat.com>