nbd: Map NbdExport names to smaller string for q-s-d
Passing the CSI Volume name (which we want exposed as our NBD export
name) to qemu-storage-daemon fails with "Node name too long" since
pvc-$UUID is longer than qemu's 32-byte blockdev name limit.
Thankfully, the export name is not as severely restricted.  But for
blockdev names, it does mean we need to map public names into shorter
internal ones for the life of the q-s-d process.

Tweak the unit test to demonstrate the problem being solved.

Signed-off-by: Eric Blake <eblake@redhat.com>