build: Fix tests/run.sh after local `make` run
In a CI environment, when nothing exists locally other than what was
in git, the containerized-build.sh script using "COPY ./ ./" is just
fine, since running "make" is going to build the world the first time.
But in local development, when ./bin has been created after "make",
when tests/run.sh tries to build a container, the "COPY ./ ./" copies
in binaries built for the host libc, but then tries to execute them in
the (often intentionally-older) libc of the container environment,
which fails if the binary depends on newer glibc symbols.

Fix this by splitting up explicit COPY lines in the Containerfile.
While at it, explicitly run a few steps earlier than the COPY so that
we can cache the results that are more likely to be slow-changing
(such as populating ./bin) while still reacting to development in the
final "RUN make build".

Someday, we may want to pre-build a container with controller-gen and
golangci-lint in that container, then run this container with "make
CONTROLLER_GEN=... GOLANGCI_LINT=..." pointing to those tools from the
imported container, to further reduce the amount of work needed in CI,
but that's not the goal of this patch.

Fixes: 37e42839 ("Makefile, Containerfile: run makefile as part of container build")
Signed-off-by: Eric Blake <eblake@redhat.com>