nbd: Improvements to NBD server manipulation
Add helper functions here that will make the upcoming changes for
adding NBD support to the Volume reconciler more concise.

In particular, my original design idea of having the Volume reconciler
create the NbdExport CR and set Volume.Status.NbdExport to the CR's
name, then to trigger deletion and set Volume.Status.NbdExport to "",
relying on kubernetes finalizers to ensure that the deletion is
delayed until clients disconnect, turned out to be hard to implement.
The client nodes were indeed waking up in a Volume reconcile() loop,
but had no easy way to see whether the NbdExport object was the reason
they woke up because the name was no longer available, so they would
have had to search the entire list of NbdExports in the namespace.

Instead, it is easier if the node running the server has a way to
trigger the NbdExport to be unavailable without doing deletion yet,
and leave Volume.Status.NbdExport set; then the nodes running clients
will have easy access to the CR during their reconcile loops.  Since
NbdExport.Spec.Path was previously only referenced to set up
NbdExport.Status.Uri, it is fairly easy to make this field nullable to
have the same effect as what previously happened when deleting the
object, except that now we don't have to worry about data races on
whether clients know which object triggered them to wake up for a
reconcile.  The server node will still eventually delete the NbdExport
when it sees all clients have disconnected.

Since the Status.Conditions["Available"] can be toggled in several
places, it is also helpful to provide human-readable reasons of what
caused the most recent toggle.

Signed-off-by: Eric Blake <eblake@redhat.com>