volume: Fix mismerge with dm activation
Commit 76a46ad4 was supposed to add a WatchPending failure to
reconcileThinAttaching(), but due to some merge conflicts encountered
while rebasing, it ended up adding the condition to
reconcileThinDetatching() instead (where the logic is wrong - when
detaching, we expect the LV to be deactivated soon).  Later, commit
671520af added the intended WatchPending check in the right place, but
without noticing that the wrong place was still intact.  Clean up the
mess now.

Fixes: 76a46ad4 ("volume: Wait for volume activation with WatchPending")
Signed-off-by: Eric Blake <eblake@redhat.com>