tests: Add helper for RWX volume creation, use it in migration
The migration test was accidentally using a RWO volume yet accessing
it from two nodes at once, ever since it was first added to the test
suite.  csi-sanity will probably insist that we catch that as an
error; but I'm more surprised that the k8s CSI sidecar didn't blink at
what was supposed to be a forbidden setup.  At any rate, rather than
open-coding a RWX volume, it's time to enhance the helper library to
make it easier to create rwo vs. rwx volumes, and update all affected
callers.

Fixes: 7c9b0e59 ("tests/t: Add test for blob pool migration")
Signed-off-by: Eric Blake <eblake@redhat.com>