crd: Give CRDs a short name, add 'kubesan' and 'lv' groups
Make it easier to type kubectl commands for investigating KubeSAN
resources, by adding abbreviations for each CRD, and by associating
each CRD with the 'kubesan' group.  Thus, even when installed in an
alternate namespace, it is now possible to do `kubectl get -A kubesan`
to list resources from all namespaces claiming to be in the kubesan
group - which happens to be only the CRDs in the kubesan-system (or
openshift-storage) namespace.  Furthermore, since both Volume and
Snapshot are logical volumes within a pool, it's nice to have an 'lv'
group.

Signed-off-by: Eric Blake <eblake@redhat.com>