deploy: Use better labels on all KubeSAN resources
Sometimes, it's nice to be able to learn which pods are part of
KubeSAN, regardless of what namespace it was installed into.  The
easiest way to do this is to add a dedicated label on all Pods created
as part of the deployment and daemonsets.

However, there's no need for our labels to live in a one-off
kubesan.gitlab.io/ prefix.  Kubernetes itself provides recommended
labels that convey the same meaning, and which are more likely to
already integrate well with existing 3rd party tools looking for
common labels; and kustomize makes it easy to apply a global label
during deployment.

In touching all uses of the old label names, I also fixed a few typos
in the testsuite utilities that were overlooked when we switched to
kubebuilder with its change to a default namespace of kubesan-system.

You can now do `kubectl get -A -l kubesan.gitlab.io/core=true pod` to
see all the pods running KubeSAN code.

(More selfishly, as a developer, I can use:
$ cat kustomization.yaml
resources:
  - deploy/openshift
images:
  - name: quay.io/kubesan/kubesan
    newName: quay.io/eblake_redhat/kubesan
    newTag: testing

patches:
  - target:
      kind: Deployment
    patch: |-
      - op: add
        path: /spec/template/spec/containers/0/imagePullPolicy
        value: Always
  - target:
      kind: DaemonSet
    patch: |-
      - op: add
        path: /spec/template/spec/containers/0/imagePullPolicy
        value: Always
$ kubectl apply -k .
$ kubectl delete -A -l app.kubernetes.io/name=kubesan pod

as a way to quickly test out the latest version of my testing image
that I just pushed to quay.io, when my $KUBECONFIG is targetting a
remote OpenShift cluster)

Signed-off-by: Eric Blake <eblake@redhat.com>