crd: Add Status.ObservedGeneration
Advertise what version of the .Spec was used to achieve the current
contents of .Status.  This can be helpful when viewing logs to search
for races, such as determining whether a reconcile event is still in
progress or may have been missed.

Note that code should not rely on just this field for performing
synchronization: k8s can batch several .Spec or .Status updates into
one, and the mere presence of this field does not tell you whether a
reconciler may try to write to .Status more than once or if competing
reconcilers overwrote one another (other patches will be more
concerned with making sure parallel reconcilers don't try to write
different status).  Thus, metadata.Generation ==
.Status.ObservedGeneration only tells you that the reconciler has seen
your latest .Spec, but may or may not be done changing .Status to
match yet.  But metadata.Generation > .Status.ObservedGeneration is a
sure sign that your requested changes have not happened yet.  And
metadata.Generation < .Status.ObservedGeneration means that you
managed to read a cached .Status that has been overwritten by someone
else, coupled with a reconciler that has already started acting on the
newer state that you would see when doing a new Get.

Fixes: https://gitlab.com/kubesan/kubesan/-/issues/77
Signed-off-by: Eric Blake <eblake@redhat.com>