tests: Add new ksan-get-condition helper
Probing the value of a CRD Status.Condtions["type"] in a test is
useful, but verbose; and it's a pain to write the nested quoting that
it would require for inlining inside of a larger ksan-poll call that
also needs quoting to be passed to a shell eval.  The existing
attempts at it only work when there is a single condition; when there
are two conditions, [*]['type','status'] in the jsonpath would produce
"Available Active True True" in the output, which isn't really what we
wanted to be grep'ing.  Thus, write a new ksan-get-condition helper
that is easier to embed into a ksan-poll call, and update all affected
tests to use it.

Signed-off-by: Eric Blake <eblake@redhat.com>