thinpoollv: Default size to volume+1%, rather than hard-coded 1G
I was running into a failure to test a Fedora image on an OpenShift
cluster, while a smaller cirros image worked fine. In the end, I
traced it to the fact that we do not yet have auto-extend working (see
#81), and while my cirros image fit in 256M, my Fedora image wanted a
disk size of 30G which was well beyond the thin pool's initial 1G
size, and thus hitting EIO errors due to the underlying ENOSPC.

This patch is not the final solution, but merely something that will
unbreak the use of KubeSAN with images larger than 1G.  As such, we
may later relax the restriction against changing the thin pool size
(even with auto-resize working, there may be times when the user wants
to resize faster than by waiting for automatic events to happen).

Signed-off-by: Eric Blake <eblake@redhat.com>