tests: Add ksan-reimage
The time spent to fire up a cluster (especially with kcli) far
outweighs the time spent testing if a code tweak worked.  While
rebuilding the cluster is sometimes necessary (depending on how much
CRDs changed, or how badly botched the last code attempt was), it is
often feasible to do further debugging by merely rebuilding the
container images and redeploying those into the existing test cluster.

Add a new utility function to be invoked during a --pause-on-stage
shell for reloading the latest code changes into the cluster.  This
also needed exposing the __log family of functions into that
environment.

There are still other things that would be nice to do in a test
environment which aren't tackled here, such as cleaning up the
existing deployment (force removing all objects in the kubesan-system
namespace, wiping all lvm objects, and redeploying) in order to
restart the current test or even swap to another test, all while still
using the current cluster.  But those will have to remain on the wish
list for another patch.

Signed-off-by: Eric Blake <eblake@redhat.com>