tests: Guarantee ksan-stage before cleanup of all tests
Commit e9503412 introduced a trailing ksan-stage on each test, with
the justification that it is very handy when developing a test to use
--pause-on-stage and check that the cluster is back to a clean state
(once the test attempts to delete the namespace, it is much harder to
read logs or depend on further cleanup items happening from
now-deleted pods).  But that commit touched lots of tests, and has the
further problem that tests that were developed in parallel lacked the
same final clause (see commits 5b2a2a6f, 77c0a102).  So let's put this
final stage in run.sh instead, so we don't have to keep doing per-test
tweaks (more or less reverting the per-test changes of e9503412).

There is one slight wrinkle, exposed by csi-sanity.sh: no test should
ever do a bare exit except with non-zero status (a bare exit bypasses
all code after the 'source file' line that ran the indivitual
tests/t/*.sh file), so there is a nice comment on why that file was
modified to instead use exit within a subshell.

Signed-off-by: Eric Blake <eblake@redhat.com>