tests: Make csi-sanity run on node with csi-controller
The cluster is not required to host the csi-controller on node 0; yet
/var/lib/kubelet/plugins/kubesan-controller must exist on the node
where the csi-sanity pod is running for the connection to succeed.
Another ksan-stage point was helpful in debugging why my test runs
would sometimes fail with no connection to the cluster controller.

This got me far enough to see that csi-sanity is currently hanging
(when I previously had it running, it would complete within 60
seconds, but that was before our CRD refactoring).  The particular
failure is that when the CSI tester tries to create a volume named
sanity-
controller-create-maxlen-775f6abd-c954373eaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa, lvcreate complains that it
is too long, but the complaint is not treated as fatal, so the test
goes into an inf-loop repeatedly trying the creation.

Signed-off-by: Eric Blake <eblake@redhat.com>